chore pt2

+++ app/models/cash_register_movement.py
@@ -1,5 +1,7 @@
 from datetime import datetime
-from sqlalchemy import Integer, String, Float, ForeignKey, DateTime, func, Text
+from decimal import Decimal
+
+from sqlalchemy import Integer, String, Numeric, ForeignKey, DateTime, func, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.core.database import Base
@@ -13,7 +15,7 @@ class CashRegisterMovement(Base):
         ForeignKey("cash_register_sessions.id"), nullable=False
     )
     type: Mapped[str] = mapped_column(String(20), nullable=False)
-    amount: Mapped[float] = mapped_column(Float, default=0.0, nullable=False)
+    amount: Mapped[Decimal] = mapped_column(Numeric(14, 2), default=0, nullable=False)
     note: Mapped[str | None] = mapped_column(Text, nullable=True)
     created_by_id: Mapped[int | None] = mapped_column(ForeignKey("users.id"), nullable=True)
     created_at: Mapped[datetime] = mapped_column(
