+++ app/models/cash_register_session.py
@@ -1,5 +1,7 @@
 from datetime import datetime
-from sqlalchemy import Integer, String, Float, ForeignKey, DateTime, func, Text
+from decimal import Decimal
+
+from sqlalchemy import Integer, String, Numeric, ForeignKey, DateTime, func, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.core.database import Base
@@ -15,8 +17,8 @@ class CashRegisterSession(Base):
     closed_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
     opened_by_id: Mapped[int] = mapped_column(ForeignKey("users.id"), nullable=False)
     closed_by_id: Mapped[int | None] = mapped_column(ForeignKey("users.id"), nullable=True)
-    initial_cash: Mapped[float] = mapped_column(Float, default=0.0, nullable=False)
-    final_cash: Mapped[float | None] = mapped_column(Float, nullable=True)
+    initial_cash: Mapped[Decimal] = mapped_column(Numeric(14, 2), default=0, nullable=False)
+    final_cash: Mapped[Decimal | None] = mapped_column(Numeric(14, 2), nullable=True)
     status: Mapped[str] = mapped_column(String(20), default="open", nullable=False)
     observations: Mapped[str | None] = mapped_column(Text, nullable=True)
     created_at: Mapped[datetime] = mapped_column(
