+++ app/models/daily_payment.py
@@ -1,5 +1,7 @@
 from datetime import datetime
-from sqlalchemy import Integer, Float, DateTime, ForeignKey, func, Text
+from decimal import Decimal
+
+from sqlalchemy import Integer, Numeric, DateTime, ForeignKey, func, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.core.database import Base
@@ -10,7 +12,7 @@ class DailyPayment(Base):
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     employee_id: Mapped[int] = mapped_column(Integer, ForeignKey("employees.id"), nullable=False)
-    amount: Mapped[float] = mapped_column(Float, nullable=False)
+    amount: Mapped[Decimal] = mapped_column(Numeric(14, 2), nullable=False)
     payment_date: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False)
     notes: Mapped[str | None] = mapped_column(Text, nullable=True)
     registered_by_id: Mapped[int] = mapped_column(Integer, ForeignKey("users.id"), nullable=False)
