+++ app/models/order_item.py
@@ -1,5 +1,7 @@
 from datetime import datetime
-from sqlalchemy import Integer, Float, ForeignKey, DateTime, Boolean, func
+from decimal import Decimal
+
+from sqlalchemy import Integer, Numeric, ForeignKey, DateTime, Boolean, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from app.core.database import Base
@@ -13,8 +15,8 @@ class OrderItem(Base):
     order_round_id: Mapped[int | None] = mapped_column(ForeignKey("order_rounds.id"), nullable=True)
     product_id: Mapped[int] = mapped_column(ForeignKey("products.id"), nullable=False)
     quantity: Mapped[int] = mapped_column(Integer, nullable=False)
-    unit_price: Mapped[float] = mapped_column(Float, nullable=False)
-    unit_cost: Mapped[float | None] = mapped_column(Float, nullable=True)
+    unit_price: Mapped[Decimal] = mapped_column(Numeric(14, 2), nullable=False)
+    unit_cost: Mapped[Decimal | None] = mapped_column(Numeric(14, 4), nullable=True)
     is_pending: Mapped[bool] = mapped_column(Boolean, default=False, nullable=False)
     created_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True), server_default=func.now()
